backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
environs==14.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastapi-cors==0.0.6
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==4.0.0
mcp==1.9.4
mdurl==0.1.2
mypy-extensions==1.1.0
openai==1.90.0
openai-agents==0.0.19
orjson==3.10.18
packaging==25.0
pathspec==0.12.1
pip==25.1.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pygments==2.19.2
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-extensions==4.14.0
typing-inspection==0.4.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
